pyproject.toml
[tool.poetry]
name = "autonomous-ml-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.12"
streamlit = "^1.26.0"
pandas = "^2.0.3"
requests = "^2.31.0"
openai = "^1.13.3"
e2b-code-interpreter = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
